requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow>=10.0.1

# OCR and Text Processing
easyocr==1.7.1
opencv-python==4.9.0.80

# NLP and Text Similarity
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1

# LLM Integration
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.30
langchain-google-genai>=0.0.6

# PDF Output
reportlab==4.1.0
fpdf2==2.7.8

# Utilities
python-dateutil==2.9.0
regex==2023.12.25
tqdm==4.66.2
click==8.1.7

# FastAPI and Uvicorn
fastapi==0.111.0
uvicorn==0.29.0 